+++ controllers/contacts/index.js
@@ -1,0 +1,17 @@
+const { ctrlWrapper } = require("../../helpers");
+
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const remove = require("./remove");
+const update = require("./update");
+const updateStatusContact = require("./updateStatusContact");
+
+module.exports = {
+  getAll: ctrlWrapper(getAll),
+  getById: ctrlWrapper(getById),
+  add: ctrlWrapper(add),
+  remove: ctrlWrapper(remove),
+  update: ctrlWrapper(update),
+  updateStatusContact: ctrlWrapper(updateStatusContact),
+};
