+++ controllers/contacts/remove.js
@@ -1,0 +1,13 @@
+const { Contact } = require("../../models/contact");
+const { HttpError } = require("../../helpers");
+
+const remove = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.status(200).json({ message: "contact deleted" });
+};
+
+module.exports = remove;
